+++ sympy/assumptions/ask.py
@@ -115,40 +115,8 @@ class AssumptionKeys:
 
     @memoize_property
     def nonpositive(self):
-        """
-        Nonpositive real number predicate.
-
-        Explanation
-        ===========
-
-        ``ask(Q.nonpositive(x))`` is true iff ``x`` belongs to the set of
-        negative numbers including zero.
-
-        - Note that ``Q.nonpositive`` and ``~Q.positive`` are *not* the same
-          thing. ``~Q.positive(x)`` simply means that ``x`` is not positive,
-          whereas ``Q.nonpositive(x)`` means that ``x`` is real and not
-          positive, i.e., ``Q.nonpositive(x)`` is logically equivalent to
-          `Q.negative(x) | Q.zero(x)``.  So for example, ``~Q.positive(I)`` is
-          true, whereas ``Q.nonpositive(I)`` is false.
-
-        Examples
-        ========
-
-        >>> from sympy import Q, ask, I
-
-        >>> ask(Q.nonpositive(-1))
-        True
-        >>> ask(Q.nonpositive(0))
-        True
-        >>> ask(Q.nonpositive(1))
-        False
-        >>> ask(Q.nonpositive(I))
-        False
-        >>> ask(Q.nonpositive(-I))
-        False
-
-        """
-        return Predicate('nonpositive')
+        from .handlers.order import NonPositivePredicate
+        return NonPositivePredicate()
 
     @memoize_property
     def nonnegative(self):
@@ -1111,7 +1079,6 @@ _handlers = [
     ("commutative",       "AskCommutativeHandler"),
     ("composite",         "ntheory.AskCompositeHandler"),
     ("even",              "ntheory.AskEvenHandler"),
-    ("nonpositive",       "order.AskNonPositiveHandler"),
     ("prime",             "ntheory.AskPrimeHandler"),
     ("odd",               "ntheory.AskOddHandler"),
     ("is_true",           "common.TautologicalHandler"),
