+++ sympy/assumptions/ask_generated.py
@@ -108,6 +108,8 @@ def get_known_facts_cnf():
         Q.hermitian | ~Q.real,
         Q.nonnegative | ~Q.positive,
         Q.nonnegative | ~Q.zero,
+        Q.nonpositive | ~Q.negative,
+        Q.nonpositive | ~Q.zero,
         Q.nonzero | ~Q.negative,
         Q.nonzero | ~Q.positive,
         Q.rational | ~Q.integer,
@@ -126,14 +128,13 @@ def get_known_facts_cnf():
         ~Q.irrational | ~Q.rational,
         ~Q.negative | ~Q.positive,
         ~Q.negative | ~Q.zero,
-        ~Q.negative | Q.nonpositive,
         ~Q.positive | ~Q.zero,
         ~Q.zero | Q.even,
-        ~Q.zero | Q.nonpositive,
         Q.positive | ~Q.prime,
         Q.invertible | Q.singular,
         Q.infinite | Q.real | ~Q.extended_real,
         Q.negative | Q.positive | ~Q.nonzero,
+        Q.negative | Q.zero | ~Q.nonpositive,
         Q.positive | Q.zero | ~Q.nonnegative,
         Q.complex_elements | ~Q.real_elements,
         Q.fullrank | ~Q.invertible,
@@ -153,7 +154,6 @@ def get_known_facts_cnf():
         Q.triangular | ~Q.upper_triangular,
         Q.unitary | ~Q.orthogonal,
         Q.upper_triangular | ~Q.diagonal,
-        Q.negative | Q.zero | ~Q.nonpositive,
         Q.negative | Q.positive | Q.zero | ~Q.real,
         ~Q.composite | ~Q.prime,
         ~Q.even | ~Q.odd,
